backend/requirements.txt
fastapi
uvicorn
pydantic[email]
beautifulsoup4
requests
langchain
langchain-google-genai
python-dotenv
google-search-results
langchain-community
